stroker: Simplify handling of varyings and fix a few bugs with it.
Rather than having the varying generator generate both last and
current point, I'll instead let the stroker cache the varyings
as part of the segment and thus let the generator only generate
the current one. This makes for a much nicer API as the generator
no longer needs to think about 'current' and 'last' and it
fixes issues with joins, caps and closing as the target
varyings are now know for each of those.