fix startcap problem introduced with eccc66e9
Some strokes use a dedicated starting length which means that relying on
Stroker::length to decide if we can place the cap or not is not a valid
criteria.

Rather, we need to check if the previous segment was a valid line
segment or not. As we're rejecting identical lineTo's, having a
line means we've moved from the start position.